fix: 启动自动同步 + qty类型改为Double
- init中自动调syncFromCloud()
- PurchaseItem.qty: Int→Double (兼容云端int/float混合)
- 所有qty显示智能格式化 (1.0→1, 1.5→1.5)
- addPurchase参数qty改为Double

+++ android/app/src/main/java/com/lucaslei/app/DataStore.kt
@@ -12,7 +12,7 @@ data class PurchaseItem(
     val name: String = "",
     val brand: String = "",
     val spec: String = "",
-    val qty: Int = 0,
+    val qty: Double = 0.0,
     val price: Double = 0.0,
     val channel: String = "",
     val contact: String = "",
